integrate browserify, switch flows
