+++ _frontapp/app.js
@@ -1,0 +1,13 @@
+/*
+ * SKG Tech
+ * Thessaloniki Tech Events and Meetups
+ * https://github.com/skgtech/skgtech.github.io
+ *
+ * Copyright (c) 2014 SKG Tech Contributors.
+ * Licensed under the MIT license.
+ */
+/**
+ * @fileOverview Frontend application bootstrap file
+ */
+
+require('./frontpage');
