+++ _includes/themes/twitter/default.html
@@ -75,11 +75,13 @@
     ================================================== -->
     <!-- Placed at the end of the document so the pages load faster -->
     <script src="https://code.jquery.com/jquery-1.10.2.min.js"></script>
+    <!--
     <script src="assets/js/chart.js"></script>
     <script src="assets/js/bootstrap.js"></script>
     <script src="assets/components/calendarth/build/calendarth.min.js"></script>
     <script src="assets/js/frontpage.js"></script>
-    <script src="https://apis.google.com/js/client.js?onload=gapiload"></script>
+    -->
+    <script src="/assets/app.js"></script>
     <script type="text/javascript">
       // twitter
       !function(d,s,id){var js,fjs=d.getElementsByTagName(s)[0],p=/^http:/.test(d.location)?'http':'https';if(!d.getElementById(id)){js=d.createElement(s);js.id=id;js.src=p+'://platform.twitter.com/widgets.js';fjs.parentNode.insertBefore(js,fjs);}}(document, 'script', 'twitter-wjs');
