+++ assets/app.js
@@ -1,0 +1,1 @@
+!function a(b,c,d){function e(g,h){if(!c[g]){if(!b[g]){var i="function"==typeof require&&require;if(!h&&i)return i(g,!0);if(f)return f(g,!0);throw new Error("Cannot find module '"+g+"'")}var j=c[g]={exports:{}};b[g][0].call(j.exports,function(a){var c=b[g][1][a];return e(c?c:a)},j,j.exports,a,b,c,d)}return c[g].exports}for(var f="function"==typeof require&&require,g=0;g<d.length;g++)e(d[g]);return e}({1:[function(a){a("./frontpage")},{"./frontpage":2}],2:[function(a,b){var c=a("calendarth"),d=b.exports={};d.init=function(){var a=c({apiKey:"AIzaSyC75rnKyEkGxmVyG7hlqFicwPBgDmQLN_w",calendarId:"2ul10sd9g30mnk1vpmcnnp5qv4@group.calendar.google.com",maxResults:6});a.fetch(function(a,b){a||console.log(b)})},d.init()},{calendarth:4}],3:[function(a,b){var c=a("./item"),d=b.exports=function(a){this.options=a||{},this.apiKey=this.options.apiKey||null,this.calendarId=this.options.calendarId||null,this.maxResults=this.options.maxResults||20};d.prototype.fetch=function(a){var b="https://www.googleapis.com/calendar/v3/calendars/";b+=this.calendarId+"/events?key=",b+=this.apiKey;var d=new Date;b+="&orderBy=startTime",b+="&singleEvents=true",b+="&timeMin="+d.toISOString(),b+="&maxResults="+this.maxResults,$.ajax({type:"GET",url:b,crossDomain:!0,dataType:"json"}).done(function(b){var d=new c(b);a(null,d)}).fail(function(b,c,d){a(d)})}},{"./item":5}],4:[function(a,b){var c=a("./calendarth");b.exports=function(a){return new c(a)}},{"./calendarth":3}],5:[function(a,b){var c=b.exports=function(a){return a};c.prototype.get=function(a){return a?this.calendarObj[a]:this.calendarObj}},{}]},{},[1]);
