+++ package.json
@@ -14,7 +14,11 @@
     "grunt-contrib-connect": "~0.7.1",
     "load-grunt-tasks": "^0.4.0",
     "grunt-contrib-cssmin": "^0.9.0",
-    "grunt-parallel": "^0.3.1"
+    "grunt-parallel": "^0.3.1",
+    "browserify": "^3.39.0",
+    "grunt-browserify": "^2.0.3",
+    "grunt-contrib-uglify": "^0.4.0",
+    "calendarth": "0.0.3"
   },
   "scripts": {},
   "repository": {
